make aws sdk a dev dependency, only build for publish
Since the createDeps allows any AWS SDK V3 version to be provided, the AWS SDK imported here was only for typescript types. Once compiled to js the types are no longer needed, so this can be a dev dependency.

Only run tests and build when publishing, not on install or pack.