change interface to better fit usage
After using the interface, it would be much easier to use in real use
cases by passing in the s3Client and only returning an object of the key
dependency functions. This makes it easy to wrap those with retry logic.

s3Client and an optional logger are passed into the CopyMultipart
constructor.